Lista de cambios: -"guardar comentario" y su funcionalidad - boton de eliminar comentario (para el usuario logueado, con el rol correcto) - Se restauro la BD del ultimo commit de Robertux, 	- agregue mi usuario 	- agregue la tabla articulos (para jwiki) 	- modifique mis tablas 	- cambie el jhard.dbschema 	- agregue la entidad articulos Y se arreglaro un talego de bugfixes, tengo bugs pero son de usabilidad.

+++ trunk/proyectos/JHard/src/java/edu/ues/jhard/beans/BeanBaseJProcur.java
@@ -512,6 +512,7 @@ public class BeanBaseJProcur extends BeanBase {
     public void deleteComentario(Integer idcomentario){
         EntityManager em = this.getEntityManager();
         Comentarios c = (Comentarios)em.find(Comentarios.class, idcomentario);
+        if(c==null) return;
         em.getTransaction().begin();
         em.remove(c);
         em.getTransaction().commit();
